add paranoidi suggestions
